+++ src/www/app/main.ts
@@ -17,9 +17,8 @@ import '../ui_components/app-root.js';
 import {EventQueue} from '../model/events';
 
 import {App} from './app';
-import {shadowsocksConfigToAccessKey} from './config';
 import {onceEnvVars} from './environment';
-import {OutlineServerFactory, OutlineServerRepository} from './outline_server';
+import {OutlineServerFactory, OutlineServerRepository, shadowsocksConfigToAccessKey} from './outline_server';
 import {OutlinePlatform} from './platform';
 import {Settings} from './settings';
