+++ src/www/app/outline_server.ts
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import {makeConfig, SHADOWSOCKS_URI, SIP002_URI} from 'ShadowsocksConfig';
 import * as uuidv4 from 'uuidv4';
 
 import * as errors from '../model/errors';
 import * as events from '../model/events';
 import {Server, ServerRepository} from '../model/server';
 
-import {accessKeyToShadowsocksConfig, ShadowsocksConfig, shadowsocksConfigToAccessKey} from './config';
+import {ShadowsocksConfig} from './config';
 import {Tunnel, TunnelStatus} from './tunnel';
 
 export class OutlineServer implements Server {
@@ -145,10 +146,6 @@ export class OutlineServerRepository implements ServerRepository {
   }
 
   add(accessKey: string, serverName: string) {
-    const alreadyAddedServer = this.serverFromAccessKey(accessKey);
-    if (alreadyAddedServer) {
-      throw new errors.ServerAlreadyAdded(alreadyAddedServer);
-    }
     const config = accessKeyToShadowsocksConfig(accessKey);
     const server = this.createServer(uuidv4(), serverName, config, this.eventQueue);
     this.serverById.set(server.id, server);
@@ -193,8 +190,26 @@ export class OutlineServerRepository implements ServerRepository {
     this.lastForgottenServer = null;
   }
 
-  containsServer(accessKey: string): Server {
-    return this.serverFromAccessKey(accessKey);
+  validateAccessKey(accessKey: string, outlineServerName?: string, defaultServerName?: string):
+      string|undefined {
+    const alreadyAddedServer = this.serverFromAccessKey(accessKey);
+    if (alreadyAddedServer) {
+      throw new errors.ServerAlreadyAdded(alreadyAddedServer);
+    }
+    let config = null;
+    try {
+      config = SHADOWSOCKS_URI.parse(accessKey);
+    } catch (error) {
+      throw new errors.ServerUrlInvalid(error.message || 'failed to parse access key');
+    }
+    if (config.host.isIPv6) {
+      throw new errors.ServerIncompatible('unsupported IPv6 host address');
+    }
+    if (!OutlineServer.isServerCipherSupported(config.method.data)) {
+      throw new errors.ShadowsocksUnsupportedCipher(config.method.data || 'unknown');
+    }
+    return config.extra?.outline && outlineServerName ? outlineServerName :
+                                                        config.tag?.data || defaultServerName;
   }
 
   private serverFromAccessKey(accessKey: string): OutlineServer|undefined {
@@ -246,6 +261,32 @@ export class OutlineServerRepository implements ServerRepository {
       }
     }
   }
+}
+
+// Parses an access key string into a ShadowsocksConfig object.
+export function accessKeyToShadowsocksConfig(accessKey: string): ShadowsocksConfig {
+  try {
+    const config = SHADOWSOCKS_URI.parse(accessKey);
+    return {
+      host: config.host.data,
+      port: config.port.data,
+      method: config.method.data,
+      password: config.password.data,
+    };
+  } catch (error) {
+    throw new errors.ServerUrlInvalid(error.message || 'failed to parse access key');
+  }
+}
+
+// Enccodes a Shadowsocks proxy configuration into an access key string.
+export function shadowsocksConfigToAccessKey(config: ShadowsocksConfig): string {
+  // Exclude the name and only encode proxying parameters.
+  return SIP002_URI.stringify(makeConfig({
+    host: config.host,
+    port: config.port,
+    method: config.method,
+    password: config.password,
+  }));
 }
 
 // Compares access keys proxying parameters.
